+++ docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 CHANGELOG
 ---------
 
+#### Next Release
+- Bug fixes:
+  - Properly handled error case when downloading & removing translations, and downloading Strong's numbers
+
 #### v0.15.0 (2020-05-22)
 - Bug fixes:
   - Used SupervisorJob for parent scope